Split tile brush controller
